+++ src/steps/resource-manager/monitor/index.test.ts
@@ -1,9 +1,26 @@
 import { Recording } from '@jupiterone/integration-sdk-testing';
 import { IntegrationConfig } from '../../../types';
-import { setupAzureRecording } from '../../../../test/helpers/recording';
+import {
+  getMatchRequestsBy,
+  setupAzureRecording,
+} from '../../../../test/helpers/recording';
 import { createMockAzureStepExecutionContext } from '../../../../test/createMockAzureStepExecutionContext';
 import { ACCOUNT_ENTITY_TYPE } from '../../active-directory';
-import { fetchLogProfiles } from '.';
+import {
+  buildActivityLogScopeRelationships,
+  fetchActivityLogAlerts,
+  fetchLogProfiles,
+} from '.';
+import { configFromEnv } from '../../../../test/integrationInstanceConfig';
+import {
+  getMockAccountEntity,
+  getMockResourceGroupEntity,
+  getMockSubscriptionEntity,
+} from '../../../../test/helpers/getMockEntity';
+import { MonitorEntities, MonitorRelationships } from './constants';
+import { Entity, getRawData } from '@jupiterone/integration-sdk-core';
+import { ActivityLogAlertResource } from '@azure/arm-monitor/esm/models';
+import { EOL_MATCHER, SUBSCRIPTION_MATCHER } from '../utils/matchers';
 
 let recording: Recording;
 
@@ -76,4 +93,160 @@ test('step = monitor log profiles', async () => {
       displayName: 'USES',
     }),
   );
+});
+
+describe('rm-monitor-activity-log-alerts', () => {
+  function getSetupEntities(config: IntegrationConfig) {
+    const accountEntity = getMockAccountEntity(config);
+    const resourceGroupEntity = getMockResourceGroupEntity('j1dev');
+
+    return { accountEntity, resourceGroupEntity };
+  }
+
+  test('success', async () => {
+    recording = setupAzureRecording({
+      directory: __dirname,
+      name: 'rm-monitor-activity-log-alerts',
+      options: {
+        matchRequestsBy: getMatchRequestsBy({ config: configFromEnv }),
+      },
+    });
+
+    const { accountEntity, resourceGroupEntity } = getSetupEntities(
+      configFromEnv,
+    );
+
+    const context = createMockAzureStepExecutionContext({
+      instanceConfig: configFromEnv,
+      entities: [resourceGroupEntity],
+      setData: {
+        [ACCOUNT_ENTITY_TYPE]: accountEntity,
+      },
+    });
+
+    await fetchActivityLogAlerts(context);
+
+    const activityLogAlertEntities = context.jobState.collectedEntities;
+
+    expect(activityLogAlertEntities.length).toBeGreaterThan(0);
+    expect(activityLogAlertEntities).toMatchGraphObjectSchema({
+      _class: MonitorEntities.ACTIVITY_LOG_ALERT._class,
+      schema: MonitorEntities.ACTIVITY_LOG_ALERT.schema,
+    });
+
+    const resourceGroupActivityLogAlertRelationships =
+      context.jobState.collectedRelationships;
+
+    expect(resourceGroupActivityLogAlertRelationships.length).toBe(
+      activityLogAlertEntities.length,
+    );
+    expect(
+      resourceGroupActivityLogAlertRelationships,
+    ).toMatchDirectRelationshipSchema({
+      schema: {
+        properties: {
+          _type: {
+            const:
+              MonitorRelationships.RESOURCE_GROUP_HAS_ACTIVITY_LOG_ALERT._type,
+          },
+        },
+      },
+    });
+  });
+});
+
+describe('rm-monitor-activity-log-alert-scope-relationships', () => {
+  async function getSetupEntities(config: IntegrationConfig) {
+    const subscriptionIdRegex = new RegExp(SUBSCRIPTION_MATCHER + EOL_MATCHER);
+
+    function isSubscriptionId(scope: string) {
+      return subscriptionIdRegex.test(scope);
+    }
+
+    function getSubscriptionIdFromActivityLogAlertEntity(e: Entity): string {
+      const alert = getRawData<ActivityLogAlertResource>(e);
+
+      for (const scope of alert?.scopes || []) {
+        if (isSubscriptionId(scope)) {
+          const subscriptionPrefixLength = '/subscriptions/'.length;
+          return scope.substring(subscriptionPrefixLength);
+        }
+      }
+      return 'subscription-id';
+    }
+
+    const accountEntity = getMockAccountEntity(config);
+    const resourceGroupEntity = getMockResourceGroupEntity('j1dev');
+
+    const context = createMockAzureStepExecutionContext({
+      instanceConfig: config,
+      entities: [resourceGroupEntity],
+      setData: {
+        [ACCOUNT_ENTITY_TYPE]: accountEntity,
+      },
+    });
+
+    await fetchActivityLogAlerts(context);
+
+    const activityLogAlertEntitiesWithSubscriptionScope = context.jobState.collectedEntities.filter(
+      (a) => {
+        const alert = getRawData<ActivityLogAlertResource>(a);
+
+        return alert?.scopes.some((s) => isSubscriptionId(s));
+      },
+    );
+
+    expect(
+      activityLogAlertEntitiesWithSubscriptionScope.length,
+    ).toBeGreaterThan(0);
+
+    const scopeSubscriptionId = getSubscriptionIdFromActivityLogAlertEntity(
+      activityLogAlertEntitiesWithSubscriptionScope[0],
+    );
+
+    const subscriptionEntity = getMockSubscriptionEntity({
+      ...config,
+      // In this case, subscriptionId is a constant from the `scope` property of an activityLogAlertEntity.
+      // In order to properly create a relationship, we'll need the target `subscriptionId` to match the `scope` from the alert.
+      subscriptionId: scopeSubscriptionId,
+    });
+
+    return {
+      accountEntity,
+      activityLogAlertEntities: activityLogAlertEntitiesWithSubscriptionScope,
+      subscriptionEntity,
+    };
+  }
+
+  test('success', async () => {
+    recording = setupAzureRecording({
+      directory: __dirname,
+      name: 'rm-monitor-activity-log-alert-scope-relationships',
+      options: {
+        matchRequestsBy: getMatchRequestsBy({ config: configFromEnv }),
+      },
+    });
+
+    const {
+      accountEntity,
+      subscriptionEntity,
+      activityLogAlertEntities,
+    } = await getSetupEntities(configFromEnv);
+
+    const context = createMockAzureStepExecutionContext({
+      instanceConfig: configFromEnv,
+      entities: [subscriptionEntity, ...activityLogAlertEntities],
+      setData: {
+        [ACCOUNT_ENTITY_TYPE]: accountEntity,
+      },
+    });
+
+    await buildActivityLogScopeRelationships(context);
+
+    const activityLogScopeRelationships =
+      context.jobState.collectedRelationships;
+
+    expect(activityLogScopeRelationships.length).toBeGreaterThan(0);
+    expect(activityLogScopeRelationships).toMatchDirectRelationshipSchema({});
+  });
 });